Разбивает ранее написанный код на модули.Из main.js вынесли все написанные ранее функции и перенесли их в файлы generate-photos.js и util.js. В util.js поместили функции для генерации чисел и случайного элемента массива и импортировали их позднее в generate-photos.js. В main.js импортировали функцию generatePhotos.
